Fixing Exception on start/loading of the plugin
The dependency between SAMLHighlighter and SamlTabController leaded
to the state, that SAMLHighlighter wanted to access SamlTabController
to identify whether the current request is a SAML request or not, but
SAMLTabController was not yet initialized. This dependency has been
removed. In order to do so, the mechanism to identify whether a
HttpRequest is a SAML message or not, has been extracted from the
SamlTabController. Likewise the SAML message decoding/encoding
functionalty has been extracted from SamlTabController.