decompress byte[0] = byte[0]
I find semantically more correct to return byte[0] when decompressing byte[0] instead
of throwing an exception. But this has the consequence that I had to add some changes
to the SAMLMessageDecode as well. A good exmaple why using try-catch for flow-control
is a bad idea. But I don't see a better way.