Removal of 'linebreaks'/'indenting' flag
This flag is missleading. It descibes as 'if line breaks should be inserted'.
But it translates to whether the XML document shall be formatted with identations
(the flag is passed to format.setIndenting(...)).

getStringOfDocument(doc, 0, false) is the same as
getStringOfDocument(doc, 0, true)

Also the indenting flag in getString(Document document, Boolean indenting, int indent)
is unnessecary and missleading.

getString(doc, false, 2) is the same as
getString(doc, true, 0)

From the documentation of 'setIndent(int indent)'
Sets the indentation. The document will not be indented if the indentation is set
to zero. Calling setIndenting will reset this value to zero (off) or the default (on).