Do not format XML after appling attacks
As we discussed, we agreed that the "Raw" checkbox was quite unintuitive
and should be eliminated.

After removing this checkbox, I discovered that one of our SAML
challenges in HackingLab could no longer be resolved (the Signature
Cloning Certificate attack).

Upon investigation, I found that after applying attacks or actions, such
as "Re-sign assertions," the corresponding method formatted the XML with
indentation, which disrupts signature validation. In previous versions,
this issue did not occur because when the "Raw" checkbox was unchecked,
the indentation was removed again before sending.

To address this, I have implemented a solution that eliminates any
indentation when attacks or actions are applied to the XML. The only
exception is for display purposes, such as in the information tab.

The challenge can now be resolved once again.